Separate geometry_utils into separate files for each class; fix import errors.
