+++ brachyutils/geometry/phantom_utils.py
@@ -1,17 +1,20 @@
-import json
-import os
+
+
 import warnings
-from glob import glob
-from pathlib import Path
-from typing import Dict, List, Literal, Optional, Union, Tuple
-from collections import defaultdict
+import os
 import numpy as np
-import SimpleITK as sitk
+from glob import glob
+from typing import Dict, List, Literal, Optional, Union, Tuple
+from pathlib import Path
 from copy import deepcopy
-# import pydicom
-from opentps.core.processing.imageProcessing.resampler3D import resampleImage3D
-from opentps.core.data import ROIContour, RTStruct
+from collections import defaultdict
+
+import nrrd
+import pydicom
+
 from opentps.core.data.images import CTImage, MRImage, ROIMask, Image3D
+from opentps.core.data import ROIContour, RTStruct
+from opentps.core.processing.imageProcessing.resampler3D import resampleImage3D
 from opentps.core.io.dicomIO import (  # writeRTDose,
     readDicomCT,
     readDicomMRI,
@@ -19,21 +22,6 @@ from opentps.core.io.dicomIO import (  # writeRTDose,
     writeDicomCT,
     writeRTStruct,
 )
-
-# Imports for brachy applicator
-from vtk import (
-    vtkCellArray,
-    vtkFillHolesFilter,
-    vtkPoints,
-    vtkPolyData,
-    vtkTransform,
-    vtkTransformPolyDataFilter,
-)
-from vtk.util import numpy_support
-from vtkmodules.vtkIOGeometry import vtkSTLReader, vtkSTLWriter
-
-import nrrd
-import pydicom
 
 class BrachyPhantom:
     r"""
@@ -99,7 +87,7 @@ class BrachyPhantom:
         self.xyz_format: bool = True
         self.anatomical_coordinate_system: Literal["LAS", "RAS", "LPS"] = "LPS"
         # Attributes for Egsphant files
-        from brachyutils.egsphant_utils import BrachyEgsphant
+        from brachyutils import BrachyEgsphant
 
         self.egsphant_obj: "BrachyEgsphant" = None
 
@@ -736,7 +724,7 @@ class BrachyPhantom:
         #prepare the phantom for egsphant conversion
         elif self.image_obj is not None:
             phantom_used_for_egsphant = deepcopy(self)
-            from brachyutils.egsphant_utils import BrachyEgsphant
+            from brachyutils import BrachyEgsphant
             if resample_egsphant_to is not None: #if we want to resample
                 if resample_phantom_base: #resample the phantom and structures that the egsphant is based on
                     phantom_used_for_egsphant.image_obj = resampleImage3D(
@@ -1336,919 +1324,3 @@ def _getExtentOfMask(mask: np.array) -> List[int]:
     #     if bound == mask.shape[2]:
     #         bound = mask.shape[2] - 1
     return boxInVoxel
-
-class BrachyApplicator:
-    r"""
-    Purpose:
-        - This class holds the information regarding the brachytherapy applicator.
-        as well as all the functions to support the necessary applicator operations.
-
-    Attributes:
-        - path:str := path to the applicator geometry file.
-        - name:str := name of the applicator, which is taken as the basename of the path.
-        - applicator_mesh := the vtk mesh of the applicator.
-        - verticies:np.array := the verticies of the applicator mesh.
-        - faces:np.array := the faces of the applicator mesh.
-        - origin:np.array := the origin of the applicator.
-        - rotation:np.array := the rotation of the applicator.
-        - material:str := the material of the applicator.
-        - density:float := the density of the applicator.
-        - normal:np.array := the normal of the applicator in the patient coordinate system. this is used for RapidBrachy only.
-
-    Functions:
-        - load_stl(pth_input:str)
-        - load_json(pth_input:str)
-        - to_dict()
-        - to_json(pth_output:str)
-    """
-
-    def __init__(
-        self,
-        pth_input_file: str,
-        material: str = None,
-        density: float = None,
-        origin: np.array = None,
-        rotation: np.array = None,
-        rotation_origin: np.array = None,
-        coordinates: np.array = None,
-        normal: np.array = None,
-        catheter_trajectory: list = None,
-    ) -> None:
-        """
-        Purpose:
-            - Initialize the Applicator object.
-        Inputs:
-            - pth_input_file (str): The path to the input file.
-            - material (str, optional): The material of the applicator. Defaults to None.
-            - density (float, optional): The density of the applicator. Defaults to None.
-            - origin (np.array, optional): The origin of the applicator in [x,y,z] . Defaults to None.
-            - rotation (np.array, optional): The rotation vector of the applicator in [w,x,y,z]. Defaults to None.
-            - rotation_origin (np.array, optional): The origin point with respect to which the rotaion vector is created.
-            - coordinates (np.array, optional): The coordinates of the applicator in patient frame. Defaults to None.
-            - normal (np.array, optional): The normal of the applicator in the patient frame. Defaults to None.
-            - catheter_trajectory: (list, optional): The list of start dwell poisition and end dwell position of the catheter inside
-            the applicator [[x,y,z,x,y,z]]. Defaults to None.
-        Outputs:
-            - Void: an applicator object is created dependeing on the inputs.
-        """
-        assert os.path.exists(
-            pth_input_file
-        ), f"input file {pth_input_file} does not exist"
-        self.path = pth_input_file
-        self.name = os.path.splitext(os.path.basename(self.path))[0]
-        self.applicator_mesh: vtkPolyData = None
-        self.verticies: np.array = None
-        self.faces: np.array = None
-        self.origin: np.array = np.array([0, 0, 0])  # [x, y, z]
-        self.rotation: np.array = np.array([0, 0, 0, 0])  # [w, x, y, z]
-        self.coordinates: np.array = np.array([0, 0, 0])  # [x, y, z]
-        self.material: str = None
-        self.density: float = None
-        self.normal: np.array = None
-        self.catheter_trajectory: np.array = None
-
-        input_extension = os.path.splitext(self.path)[1]
-        if input_extension == ".stl":
-            self.load_stl(self.path)
-        elif input_extension == ".json":
-            self.load_json(self.path)
-        else:
-            raise ValueError("invalid input file extension")
-
-        if material is not None:
-            self.material = material
-        if density is not None:
-            self.density = density
-        if origin is not None:
-            self.set_origin(origin)
-        if rotation is not None and rotation_origin is not None:
-            self.set_rotation(rotation, rotation_origin)
-        if coordinates is not None:
-            self.set_coordinates(coordinates)
-        if normal is not None:
-            self.normal = normal
-        if catheter_trajectory is not None:
-            self.catheter_trajectory = catheter_trajectory
-
-    def load_stl(self, pth_input: str) -> None:
-        r"""
-        Purpose:
-            - To load the applicator geometry from an stl file.
-        Inputs:
-            - pth_input:str := path to the stl file containing the applicator geometry.
-        Outputs:
-            - Void := will update the BrachyApplicator object based on the stl file.
-        """
-        reader = vtkSTLReader()
-        reader.SetFileName(pth_input)
-        reader.Update()
-        self.applicator_mesh = reader.GetOutput()
-        self._update_brachy_applicator_from_applicator_mesh()
-
-    def load_json(self, pth_input: str) -> None:
-        r"""
-        Purpose:
-            - To load the applicator geometry from a json file.
-        Inputs:
-            - pth_input:str := path to the stl file containing the applicator geometry.
-        Outputs:
-            - Void := will update the BrachyApplicator object based on the json file.
-        """
-        with open(pth_input, "r") as json_file:
-            applicator_dict = json.load(json_file)
-
-        self.verticies = np.array(applicator_dict["verticies"], dtype=np.float32)
-        self.faces = np.array(applicator_dict["faces"], dtype=np.int32)
-        self.set_origin(np.array(applicator_dict["origin"]))
-        self.set_rotation(np.array(applicator_dict["rotation"]))
-        self.set_coordinates(np.array(applicator_dict["coordinates"]))
-        self.material = applicator_dict["material"]
-        self.density = applicator_dict["density"]
-
-    def load_mac(self, pth_input: str) -> None:
-        r"""
-        Purpose:
-            - To load the applicator geometry from a mac file.
-        Inputs:
-            - pth_input:str := path to the mac file containing the applicator geometry.
-        Outputs:
-            - Void := will update the BrachyApplicator object based on the mac file.
-        """
-        raise NotImplementedError("to be implemented soon")
-
-    def info(self) -> None:
-        r"""
-        Purpose:
-            - To print the information about the applicator.
-        Inputs:
-            - self := the BrachyApplicator object.
-        Outputs:
-            - Void := will print the information about the applicator.
-        """
-        print("Applicator info is as follows:")
-        print(self.to_dict())
-
-    def is_equal(self, other) -> bool:
-        r"""
-        Purpose:
-            - To compare the current applicator with another applicator.
-        Inputs:
-            - other:BrachyApplicator := the other applicator to compare with.
-        Outputs:
-            - bool := True if the two applicators are equal, False otherwise.
-        """
-        if type(self) is not type(other):
-            return False
-        if self.name != other.name:
-            return False
-        if not np.isclose(self.verticies, other.verticies, atol=1e-6).all():
-            return False
-        if not np.isclose(self.faces, other.faces, atol=1e-6).all():
-            return False
-        if not np.isclose(self.origin, other.origin, atol=1e-6).all():
-            return False
-        if not np.isclose(self.rotation, other.rotation, atol=1e-6).all():
-            return False
-        if self.material != other.material:
-            return False
-        if self.density != other.density:
-            return False
-        return True
-
-    def _update_applicator_mesh_from_brachy_applicator(self) -> None:
-        r"""
-        Purpose:
-            - To update the applicator mesh from the verticies and faces.
-        Inputs:
-            - self := the BrachyApplicator object.
-        Outputs:
-            - Void := will update the applicator mesh from the verticies and faces.
-        """
-        points = vtkPoints()
-        for vertex in self.verticies:
-            points.InsertNextPoint(vertex)
-        self.applicator_mesh.SetPoints(points)
-
-        cell_array = vtkCellArray()
-        for face in self.faces:
-            cell_array.InsertNextCell(3, face)
-        self.applicator_mesh.SetPolys(cell_array)
-        fill_holes_filter = vtkFillHolesFilter()
-        fill_holes_filter.SetInputData(self.applicator_mesh)
-        fill_holes_filter.Update()
-        self.applicator_mesh = fill_holes_filter.GetOutput()
-
-    def _update_brachy_applicator_from_applicator_mesh(self) -> None:
-        r"""
-        Purpose:
-            - To update the brachy applicator from the applicator mesh.
-        Inputs:
-            - self := the BrachyApplicator object.
-        Outputs:
-            - Void := will update the brachy applicator from the applicator mesh.
-        """
-        self.verticies = numpy_support.vtk_to_numpy(
-            self.applicator_mesh.GetPoints().GetData()
-        )
-        self.faces = numpy_support.vtk_to_numpy(
-            self.applicator_mesh.GetPolys().GetData()
-        )
-        self.faces = self.faces.reshape(-1, 4)[:, 1:]
-
-    def set_origin(self, origin: np.array) -> None:
-        r"""
-        Purpose:
-            - To set the origin of the applicator.
-        Inputs:
-            - origin:np.array := the origin of the applicator.
-        Outputs:
-            - Void := will update the applicator verticies based on the new origin.
-        """
-        old_origin = self.origin
-        change_in_origin = np.ones_like(self.verticies) * (origin - old_origin)
-        self.origin = origin
-        self.verticies += change_in_origin
-        self._update_applicator_mesh_from_brachy_applicator()
-
-    def set_rotation(
-        self, rotation: np.array, rotation_origin: np.array = None
-    ) -> None:
-        r"""
-        Purpose:
-            - To set the rotation of the applicator.
-            the rotation origin is assumed to be the origin of applicator. To rotate the
-            applicator around its center, coordinates of the center of applicator should
-            be provided. The rotation angle is the first element of the rotation vector. the rotation
-            axis is the last three elements of the rotation vector [w,x,y,z].
-        Inputs:
-            - rotation:np.array := the rotation of the applicator.
-            The rotation vector is in quaternion ([w, x, y, z]).
-            - rotation_origin:np.array := the origin of the rotation. if not provided, the
-            origin of the applicator will be used.
-        Outputs:
-            - Void := will update the applicator verticies based on the new rotation.
-        """
-        # set the rotation attribute
-        self.rotation = rotation
-        # by default, the rotation origin is the origin of the applicator
-        # if rotation is provided, the applicator is translated to the rotation origin
-        # then it is rotated and translated back to the original position.
-        if rotation_origin is not None:
-            transform_translate = vtkTransform()
-            transform_translate.Translate(
-                -rotation_origin[0], -rotation_origin[1], -rotation_origin[2]
-            )
-            transform_translate_filter = vtkTransformPolyDataFilter()
-            transform_translate_filter.SetTransform(transform_translate)
-            transform_translate_filter.SetInputData(self.applicator_mesh)
-            transform_translate_filter.Update()
-            self.applicator_mesh = transform_translate_filter.GetOutput()
-
-        # # now apply the rotation
-        # create the transformation matrix
-        transform = vtkTransform()
-        transform.RotateWXYZ(rotation[0], rotation[1], rotation[2], rotation[3])
-
-        # apply the transformation
-        transform_filter = vtkTransformPolyDataFilter()
-        transform_filter.SetTransform(transform)
-        transform_filter.SetInputData(self.applicator_mesh)
-        transform_filter.Update()
-        self.applicator_mesh = transform_filter.GetOutput()
-
-        # if rotation origin is provided, translate the applicator back to the original position
-        if rotation_origin is not None:
-            transform_translate = vtkTransform()
-            transform_translate.Translate(
-                rotation_origin[0], rotation_origin[1], rotation_origin[2]
-            )
-            transform_translate_filter = vtkTransformPolyDataFilter()
-            transform_translate_filter.SetTransform(transform_translate)
-            transform_translate_filter.SetInputData(self.applicator_mesh)
-            transform_translate_filter.Update()
-            self.applicator_mesh = transform_translate_filter.GetOutput()
-
-        # update the BrachyApplicator based on the transformation
-        self._update_brachy_applicator_from_applicator_mesh()
-
-    def set_coordinates(self, coordinates: np.array) -> None:
-        r"""
-        Purpose:
-            - to located the applicator at a given coordinate with respect to
-            self.origin.
-        Inputs:
-            - coordinates:np.array := the coordinates of the applicator.
-        Outputs:
-            - Void := will update the applicator verticies based on the new coordinates.
-        """
-        # set the coordinate attributes
-        self.coordinates = coordinates
-
-        # create transformation matrix
-        transform = vtkTransform()
-        transform.Translate(coordinates[0], coordinates[1], coordinates[2])
-
-        # apply the transformation
-        transform_filter = vtkTransformPolyDataFilter()
-        transform_filter.SetTransform(transform)
-        transform_filter.SetInputData(self.applicator_mesh)
-        transform_filter.Update()
-        self.applicator_mesh = transform_filter.GetOutput()
-
-        # update the BrachyApplicator based on the transformation
-        self._update_brachy_applicator_from_applicator_mesh()
-
-    def _update_catheter_trajectory(
-        self,
-    ) -> None:
-        r"""
-        Purpose:
-            - to update the trajectory of the dwell positions inside the applicator after the applicator has
-            been rotated or translated.
-        Inputs:
-            - self := the BrachyApplicator object.
-        Outputs:
-            - Void := will update the catheter trajectory.
-        """
-
-        raise NotImplementedError("to be implemented soon")
-
-    def to_dict(self) -> dict:
-        r"""
-        Purpose:
-            - To convert the applicator geometry to a dictionary.
-        Inputs:
-            - self := the BrachyApplicator object.
-        Outputs:
-            - dict := the dictionary containing the applicator geometry.
-        """
-        return {
-            "name": self.name,
-            "path": self.path,
-            # "verticies": self.verticies.tolist(),
-            # "faces": self.faces.tolist(),
-            "origin": self.origin,
-            "rotation": self.rotation,
-            "material": self.material,
-            "density": self.density,
-            "normal": self.normal,
-            "catheter_trajectory": self.catheter_trajectory,
-        }
-
-    def to_json(self, pth_output: str) -> None:
-        r"""
-        Purpose:
-            - To save the applicator geometry to a json file.
-        Inputs:
-            - pth_output:str := path to the output json file.
-        Outputs:
-            - Void := will save the applicator geometry to a json file.
-        """
-        applicator_dict = self.to_dict()
-
-        with open(pth_output, "w") as json_file:
-            json.dump(applicator_dict, json_file, indent=4)
-
-    def to_mac(self, pth_output: str) -> None:
-        r"""
-        Purpose:
-            - To save the applicator geometry to a mac file.
-        Inputs:
-            - pth_output:str := path to the output mac file.
-        Outputs:
-            - Void := will save the applicator geometry to a mac file.
-        """
-        macfile_string = ""
-
-        # add in the vertex info
-        float_formatter = "{:.3f}".format
-        for vertex in self.verticies:
-            macfile_string += f"/applicator/vertex {float_formatter(vertex[0])} {float_formatter(vertex[1])} {float_formatter(vertex[2])} mm\n"
-
-        # add in the face info
-        for face in self.faces:
-            macfile_string += f"/applicator/face {face[0]} {face[1]} {face[2]}\n"
-        # add in the material info
-        macfile_string += f"/applicator/material {self.material}\n"
-        # add in the density info
-        macfile_string += f"/applicator/density {self.density}\n"
-        # add in the origin info
-        macfile_string += "/applicator/xPosition 0 mm\n"
-        macfile_string += "/applicator/yPosition 0 mm\n"
-        macfile_string += "/applicator/zPosition 0 mm\n"
-        # add in rotation nfo
-        macfile_string += "/applicator/xRotation 0 deg\n"
-        macfile_string += "/applicator/yRotation 0 deg\n"
-        macfile_string += "/applicator/zRotation 0 deg\n"
-        # add in the done flag
-        macfile_string += "/applicator/done\n"
-
-        with open(pth_output, "w") as mac_file:
-            mac_file.write(macfile_string)
-
-    def to_stl(self, pth_output: str) -> None:
-        r"""
-        Purpose:
-            - To save the applicator geometry to an stl file.
-        Inputs:
-            - pth_output:str := path to the output stl file.
-        Outputs:
-            - Void := will save the applicator geometry to an stl file.
-        """
-        self._update_applicator_mesh_from_brachy_applicator()
-        # write the polydata to an stl file
-        stl_writer = vtkSTLWriter()
-        stl_writer.SetFileName(pth_output)
-        stl_writer.SetInputData(self.applicator_mesh)
-        stl_writer.Write()
-
-
-class DwellPosition:
-    r"""
-    Purpose:
-        - This class holds the information regarding a dwell position.
-
-    Attributes:
-        - angle := angle of the IMBT shield
-        - position:dict: np.array := dwell position in the patient coordinate system [x, y, z]
-        - relativePos: int := dwell coordinate along the catheter from the reference point. increments of 5 mm
-        - rotation: np.array := rotation of the dwell position in the patient coordinate system [x, y, z]
-        - time: float := dwell time for this dwell position
-        - weight: float := ratio of this dwell time over the sum of all dwell times in all catheters.
-    """
-
-    def __init__(
-        self,
-        index: int = None,
-        angle: float = 0,
-        position: np.array = None,
-        relativePos: int = None,
-        rotation: np.array = None,
-        time: float = None,
-        weight: float = None,
-        dwell_dict: dict = None,
-    ) -> None:
-        r"""
-        Purpose:
-            - Initialize the DwellPosition object.
-        Inputs:
-            - index:int := the index of the dwell position.
-            - angle:float := angle of the IMBT shield
-            - position:np.array := dwell position in the patient coordinate system [x, y, z]
-            - relativePos:int := dwell coordinate along the catheter from the reference point. increments of 5 mm
-            - rotation:np.array := rotation of the dwell position in the patient coordinate system [x, y, z]
-            - time:float := dwell time for this dwell position
-            - weight:float := ratio of this dwell time over the sum of all dwell times in all catheters.
-            - dwell_dict:dict := the dictionary containing the dwell position.
-            either provide the index, angle, position, relativePos, rotation, time and weight or provide the dwell_dict. Not both.
-        """
-        assert (
-            (index is not None)
-            and (angle is not None)
-            and (position is not None)
-            and (relativePos is not None)
-            and (rotation is not None)
-            and (time is not None)
-            and (weight is not None)
-        ) != (
-            dwell_dict is not None
-        ), "Either provide index, angle, position, relativePos, rotation, time and weight or provide catheter_dict. Not both."
-
-        if dwell_dict is not None:
-            index = dwell_dict.get("index", None)
-            angle = float(dwell_dict.get("angle"))
-            position = np.array(
-                [
-                    dwell_dict.get("position").get("x"),
-                    dwell_dict.get("position").get("y"),
-                    dwell_dict.get("position").get("z"),
-                ]
-            )
-            relativePos = dwell_dict.get("relativePos")
-            rotation = np.array(
-                [
-                    dwell_dict.get("rotation").get("x"),
-                    dwell_dict.get("rotation").get("y"),
-                    dwell_dict.get("rotation").get("z")
-                ]
-            )
-            time = float(dwell_dict.get("time"))
-            weight = float(dwell_dict.get("weight", None))
-
-        assert isinstance(index, int), "index should be an integer"
-        self.index = index
-        assert isinstance(
-            angle, float or int
-        ), "index should be a floating point number"
-        self.angle = angle
-        assert isinstance(position, np.ndarray), "position should be a numpy array"
-        self.position = position
-        assert isinstance(relativePos, int), "relativePos should be an integer"
-        self.relativePos = relativePos
-        assert isinstance(rotation, np.ndarray), "rotation should be a numpy array"
-        self.rotation = rotation
-        assert isinstance(time, float), "time should be a float"
-        self.time = time
-        assert isinstance(weight, float), "weight should be a float"
-        self.weight = weight
-
-    def to_dict(self) -> dict:
-        r"""
-        Purpose:
-            - To convert the dwell position to a dictionary.
-        Inputs:
-            - self := the DwellPosition object.
-        Outputs:
-            - dict := the dictionary containing the dwell position.
-        """
-        return {
-            "index": int(self.index),
-            "angle": float(self.angle),
-            "position": {
-                "x": float(self.position[0]),
-                "y": float(self.position[1]),
-                "z": float(self.position[2]),
-            },
-            "relativePos": int(self.relativePos),
-            "rotation": {
-                "x": float(self.rotation[0]),
-                "y": float(self.rotation[1]),
-                "z": float(self.rotation[2]),
-            },
-            "time": float(self.time),
-            "weight": float(self.weight),
-        }
-
-
-class Catheter:
-    r"""
-    Purpose:
-        - This class holds the information regarding a catheter.
-    Attributes:
-        - id:int := the id of the catheter.
-        - points:List[np.array] := the list of points of the catheter.
-        - dwells:List[DwellPosition] := the list of dwell positions of the catheter.
-    """
-
-    def __init__(
-        self,
-        iD: int = None,
-        dwells: list = None,
-        points: List[DwellPosition] = None,
-        channel_total_time: float = None,
-        catheter_dict: dict = None,
-    ) -> None:
-        r"""
-        Purpose:
-            - Initialize the Catheter object.
-        Inputs:
-            - iD:int := the id of the catheter.
-            - dwells:List[DwellPosition] := the list of dwell positions of the catheter.
-            - points:List[np.array] := the list of points of the catheter.
-            - catheter_dict:dict := the dictionary containing the catheter.
-        """
-        assert (
-            iD is not None
-            and dwells is not None
-            and points is not None
-            and channel_total_time is not None
-        ) != (
-            catheter_dict is not None
-        ), "Either provide iD, dwells and points or provide catheter_dict. Not both."
-        if catheter_dict is not None:
-            iD = catheter_dict.get("id")
-            points = catheter_dict.get("points")
-            dwells = []
-            channel_total_time = catheter_dict.get("channel_total_time", 0.0)
-            for i, dwell_dict in enumerate(catheter_dict.get("dwells")):
-                if "index" not in dwell_dict:
-                    dwell_dict["index"] = i
-                dwells.append(DwellPosition(dwell_dict=dwell_dict))
-                if "channel_total_time" not in catheter_dict:
-                    channel_total_time += dwell_dict.get("time")
-
-        assert isinstance(iD, int), "iD should be an integer"
-        self.id = iD
-        assert isinstance(points, list), "points should be a list"
-        self.points = points
-        assert isinstance(dwells, list), "dwells should be a list"
-        self.dwells = dwells
-        assert isinstance(
-            channel_total_time, float
-        ), "channel_total_time should be a float"
-        self.channel_total_time = channel_total_time
-
-    def to_dict(self) -> dict:
-        r"""
-        Purpose:
-            - To convert the catheter to a dictionary.
-        Inputs:
-            - self := the Catheter object.
-        Outputs:
-            - dict := the dictionary containing the catheter.
-        """
-        return {
-            "id": self.id,
-            "points": self.points,
-            "dwells": [dwell.to_dict() for dwell in self.dwells],
-            "channel_total_time": self.channel_total_time,
-        }
-
-
-class CatheterTable:
-    r"""
-    Purpose:
-        - This class holds the information regarding the catheter table.
-    Attributes:
-        - catheter_list : List[Catheter] := the list of catheter objects in the catheter table.
-    Functions:
-        - load_from_json(pth_json:Path) -> list
-        - load_from_dicom(pth_dicom:Path) -> list
-    """
-
-    def __init__(
-        self,
-        catheter_list: List[Union[Catheter, dict]] = None,
-        pth_catheter_table: Path = None,
-    ) -> None:
-        r"""
-        Purpose:
-            - Initialize the CatheterTable object. from a list or a file. please provide only one of the inputs.
-        Inputs:
-            - catheter_list:List[Catheter] := the list of catheters in the catheter table.
-            - pth_catheter_table:Path := the path to the catheter table file, which could be
-            a dicom plan or a json file.
-        """
-        assert (catheter_list is not None) != (
-            pth_catheter_table is not None
-        ), "Either the catheter list or the path to the catheter table should be provided."
-
-        if pth_catheter_table is not None:
-            assert os.path.exists(
-                pth_catheter_table
-            ), f"The input json file does not exist: {pth_catheter_table}"
-            extension = os.path.splitext(pth_catheter_table)[1]
-            if extension == ".json":
-                catheter_list = self.load_from_json(pth_catheter_table)
-            elif extension == ".dcm":
-                catheter_list = self.load_from_dicom(pth_catheter_table)
-        if isinstance(catheter_list[0], dict):
-            catheter_list = [
-                Catheter(catheter_dict=catheter_dict) for catheter_dict in catheter_list
-            ]
-
-        assert isinstance(
-            catheter_list[0], Catheter
-        ), "The catheter list should contain Catheter objects."
-        self.catheter_list: list = catheter_list
-
-    def __iter__(self):
-        for catheter in self.catheter_list:
-            yield catheter
-
-    def load_from_json(self, pth_json: Path) -> list:
-        r"""
-        Purpose:
-            - Load the catheter table from a json file.
-        Inputs:
-            - pth_json: Path := the path to the json file containing the catheter table.
-        Outputs:
-            - Void := will update the catheter table based on the json file.
-        """
-        raw_catheter_table: list = []
-        with open(pth_json, "r") as json_file:
-            catheter_table_list = json.load(json_file)
-            assert isinstance(
-                catheter_table_list, list
-            ), "The json file, should contain a list of catheters."
-            for catheter_dict in catheter_table_list:
-                raw_catheter_table.append(Catheter(catheter_dict=catheter_dict))
-            return raw_catheter_table
-
-    def load_from_dicom(self, pth_dicom: Path) -> list:
-        r"""
-        Purpose:
-            - Load the catheter table from a dicom file.
-        Inputs:
-            - pth_dicom: Path := the path to the dicom file containing the catheter table.
-        Outputs:
-            - Void := will update the catheter table based on the dicom file.
-        """
-        import pydicom
-
-        plan = pydicom.dcmread(pth_dicom)
-        catheter_table = []
-        # loop through the channels
-        for catheter_dcm in plan.ApplicationSetupSequence[0].ChannelSequence:
-            control_points = []
-            catheter_time = (
-                float(catheter_dcm.ChannelTotalTime)
-                if hasattr(catheter_dcm, "ChannelTotalTime")
-                else 0
-            )
-            channel_final_time_weight = (
-                float(catheter_dcm.FinalCumulativeTimeWeight)
-                if hasattr(catheter_dcm, "FinalCumulativeTimeWeight")
-                else 0
-            )
-            # loop through the control points.
-            # Each dwell position has 2 control points, get them all.
-            for control_point_dcm in catheter_dcm.BrachyControlPointSequence:
-                if control_point_dcm.CumulativeTimeWeight is None:
-                    continue
-
-                cumulative_time_weight = (
-                    float(control_point_dcm.CumulativeTimeWeight)
-                    if hasattr(control_point_dcm, "CumulativeTimeWeight")
-                    else 0
-                )
-                control_points.append(
-                    {
-                        "index": (
-                            int(control_point_dcm.ControlPointIndex)
-                            if hasattr(control_point_dcm, "ControlPointIndex")
-                            else None
-                        ),
-                        "angle": (
-                            control_point_dcm.ControlPointShieldAngle
-                            if hasattr(control_point_dcm, "ControlPointShieldAngle")
-                            else 0
-                        ),
-                        "position": (
-                            np.array(
-                                control_point_dcm.ControlPoint3DPosition,
-                                dtype=np.float32,
-                            )
-                            if hasattr(control_point_dcm, "ControlPoint3DPosition")
-                            else None
-                        ),
-                        "relativePos": (
-                            float(control_point_dcm.ControlPointRelativePosition)
-                            if hasattr(
-                                control_point_dcm, "ControlPointRelativePosition"
-                            )
-                            else None
-                        ),
-                        "rotation": (
-                            np.array(
-                                control_point_dcm.ControlPointOrientation,
-                                dtype=np.float32,
-                            )
-                            if hasattr(control_point_dcm, "ControlPointOrientation")
-                            else np.array([0, 0, 0], dtype=np.float32)
-                        ),
-                        "cumulative_weight": cumulative_time_weight,
-                        # "total rerence air kerma": total_reference_air_kerma,
-                    }
-                )
-            catheter_table.append(
-                {
-                    "id": int(catheter_dcm.ChannelNumber) - 1,
-                    "points": [],
-                    "channel_total_time": catheter_time,
-                    "channel_final_time_weight": channel_final_time_weight,
-                    "control_points": control_points,
-                }
-            )
-
-        # # Convert control points to dwell positions:
-        # # after extracting the final cummulative time weight of the catheters,
-        # # the time of the catheter, and the cummulative time weight of the control points,
-        # # we need to calculate the dwell time and time weight of the dwell positions.
-        # # the formula is:
-        # #     time_weight = (cumulative_time_weight - previous_cumulative_time_weight) / channel_final_time_weight
-        # #     dwell time = time_weight * channel_total_time
-        # #     dwell weight = dwell time / sum(channel_total_time)
-        # get total treatment time
-        treatment_time = np.sum(
-            [catheter["channel_total_time"] for catheter in catheter_table]
-        )
-        final_catheter_table = []
-        # loop through the catheters
-        for catheter in catheter_table:
-            dwells = []
-            # loop through the control points
-            # each dwell position has 2 control points:
-            #   arrive time and depart time for the source
-            for idx, control_point in enumerate(catheter["control_points"]):
-                # if idx == len(catheter["control_points"]) - 1:
-                #     break
-                if idx % 2 == 1:
-                    continue
-                dwell_time_weight = (
-                    catheter["control_points"][idx + 1]["cumulative_weight"]
-                    - control_point["cumulative_weight"]
-                ) / catheter["channel_final_time_weight"]
-                dwell_time = dwell_time_weight * catheter["channel_total_time"]
-                dwell_weight = dwell_time / treatment_time
-                dwells.append(
-                    {
-                        "index": int(control_point["index"] / 2),
-                        "angle": float(control_point["angle"]),
-                        "position": {
-                            "x":control_point["position"][0],
-                            "y":control_point["position"][1],
-                            "z":control_point["position"][2]
-                            },
-                        "relativePos": int(control_point["relativePos"]),
-                        "rotation": {
-                            "x":control_point["rotation"][0],
-                            "y":control_point["rotation"][1],
-                            "z":control_point["rotation"][2]
-                            },
-                        "time": dwell_time,
-                        "weight": dwell_weight,
-                    }
-                )
-            catheter["dwells"] = dwells
-            if (
-                np.all([np.all(list(catheter["dwells"][i]["rotation"].values()) == [0,0,0])
-                        for i in range(len(catheter["dwells"]))])
-                and len(catheter["dwells"]) > 1
-            ):
-                for i in range(len(dwells)):
-                    dwells[i]["rotation"] = _get_rotation_from_position(i, dwells)
-    
-            final_catheter_table.append(Catheter(catheter_dict=catheter))
-        return final_catheter_table
-
-    def get_treatment_time(self) -> float:
-        r"""
-        Purpose:
-            - To calculate the total treatment time.
-        Inputs:
-            - catheter_table:CatheterTable := the catheter table object.
-        Outputs:
-            - float := the total treatment time.
-        """
-        return np.sum([catheter.channel_total_time for catheter in self.catheter_list])
-
-    def to_dict(self) -> dict:
-        r"""
-        Purpose:
-            - To convert the catheter table to a dictionary.
-        Inputs:
-            - self := the CatheterTable object.
-        Outputs:
-            - dict := the dictionary containing the catheter table.
-        """
-        return [catheter.to_dict() for catheter in self.catheter_list]
-
-    def info(self) -> None:
-        r"""
-        Purpose:
-            - To print the information about the catheter table.
-        """
-        # print(self.to_dict())
-        print("Catheter table info is as follows:")
-        print(f"Number of catheters: {len(self.catheter_list)}")
-        for catheter in self.catheter_list:
-            print(f"Catheter ID: {catheter.id}")
-            print(f"Number of dwell positions: {len(catheter.dwells)}")
-            print(f"Total channel time: {catheter.channel_total_time}")
-
-def _get_rotation_from_position(idx, control_points):
-    r"""
-    Purpose:
-        - To get the rotation of the dwell point from the position of the dwell point.
-    Inputs:
-        - idx:int := the index of the dwell point.
-        - control_point_dcm:pydicom.dataset.Dataset := the control point object.
-    Outputs:
-        - np.array := the rotation of the dwell point in each axis.
-    """
-    # TODO: Merge this dicom utils script with my catheter setup class.
-    # We need all dwell positions, not only the non 0s ones to be able to 
-    # compute correct angles when they are not provided by the DICOM.
-    if len(control_points) == 2:
-        return _angle_betwen_2_points(
-            np.array(list(control_points[1]["position"].values()), dtype=np.float32),
-            np.array(list(control_points[0]["position"].values()), dtype=np.float32),
-        )
-
-    if idx == 0:
-        return _get_rotation_from_position(idx+1, control_points)
-    elif idx == len(control_points) - 1:
-        return _get_rotation_from_position(idx-1, control_points)
-    else:
-        return _angle_betwen_2_points(
-            np.array(list(control_points[idx-1]["position"].values()), dtype=np.float32),
-            np.array(list(control_points[idx+1]["position"].values()), dtype=np.float32),
-        )
-
-
-def _angle_betwen_2_points(a, b) -> dict:
-    r"""
-    Purpose:
-        - To calculate the angle between two points.
-    Inputs:
-        - a:np.array := the first point.
-        - b:np.array := the second point.      
-    Outputs:
-        - np.array := the angle between the two points in each axis.
-    """
-    vec = a - b
-    normal = np.sqrt(np.sum(vec ** 2))
-    angle_np = vec / normal
-    return {"x":angle_np[0], "y":angle_np[1], "z":angle_np[2]}
-
